requirements.txt
pillow
pydub
